reactive fixes
