+++ src/main/java/playground/App.java
@@ -1,14 +1,17 @@
 package playground;
 
 
-import javax.annotation.PostConstruct;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.data.mongodb.repository.config.EnableReactiveMongoRepositories;
 import org.springframework.web.reactive.config.EnableWebFlux;
 import playground.config.UnirestMapperConfig;
 
+import javax.annotation.PostConstruct;
+
 @SpringBootApplication
 @EnableWebFlux
+@EnableReactiveMongoRepositories("playground.repo")
 public class App {
 
 //    @Override
@@ -16,13 +19,13 @@ public class App {
 //        return application.sources(App.class);
 //    }
 
-  public static void main(String[] args) {
-    SpringApplication.run(App.class, args);
-  }
+    public static void main(String[] args) {
+        SpringApplication.run(App.class, args);
+    }
 
-  @PostConstruct
-  public void setupUnirest() {
-    UnirestMapperConfig.setup();
-  }
+    @PostConstruct
+    public void setupUnirest() {
+        UnirestMapperConfig.setup();
+    }
 
 }
